+++ FinalProject/MapPoint/WellPoint.cs
@@ -11,7 +11,7 @@ namespace FinalProject.MapPoint
         public WellPoint()
         {
             int pointdirection;//유닛에게 전달할 디렉션번호
-            setPosition(new List<int> { 6, 0, 6 });
+            setPosition(new List<double> { 6, 0, 6 });
             setDirection(4);
             setPointType("Center");
             pointdirection = getDirection();
